backend/requirements.txt
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg[binary]
